image_filter/requirements.txt
Pillow == 8.1.0

numpy == 1.19.2

pytesseract == 0.3.7

requests == 2.25.1

tensorflow == 2.4.1

pyspellchecker == 0.6.2
